+++ demisto_sdk/commands/common/constants.py
@@ -597,8 +597,6 @@ TRIGGER_JSON_REGEX = rf"{TRIGGER_DIR_REGEX}\/([^/]+)\.json"
 
 XDRC_TEMPLATE_DIR_REGEX = rf"{PACK_DIR_REGEX}\/{XDRC_TEMPLATE_DIR}"
 XDRC_TEMPLATE_JSON_REGEX = rf"{XDRC_TEMPLATE_DIR_REGEX}\/([^/]+)\.json"
-
-XDRC_TEMPLATE_DIR_REGEX = rf"{PACK_DIR_REGEX}\/{XDRC_TEMPLATE_DIR}"
 XDRC_TEMPLATE_YML_REGEX = rf"{XDRC_TEMPLATE_DIR_REGEX}\/([^/]+)\.yml"
 
 CORRELATION_RULES_DIR_REGEX = rf"{PACK_DIR_REGEX}/{CORRELATION_RULES_DIR}"
@@ -1301,9 +1299,7 @@ FILETYPE_TO_DEFAULT_FROMVERSION = {
     FileType.GENERIC_DEFINITION: "6.5.0",
     FileType.CORRELATION_RULE: "6.10.0",
     FileType.PARSING_RULE: "6.10.0",
-    FileType.XSIAM_REPORT: "6.10.0",
     FileType.MODELING_RULE: "6.10.0",
-    FileType.XSIAM_DASHBOARD: "6.10.0",
 }
 # This constant below should always be two versions before the latest server version
 GENERAL_DEFAULT_FROMVERSION = "6.5.0"
